backend/requirements.txt
Django==4.2.3
djangorestframework==3.14.0
PyYAML==6.0
requests==2.31.0
django-cors-headers==4.2.0
